fallback: extract into crate
For faster incremental builds and to enable subsequent extraction.

Closes #12183
