+++ crates/fallback/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "fish-common"
+name = "fish-fallback"
 edition.workspace = true
 rust-version.workspace = true
 version = "0.0.0"
@@ -7,8 +7,16 @@ repository.workspace = true
 license.workspace = true
 
 [dependencies]
+fish-common.workspace = true
+fish-wchar.workspace = true
+fish-widecharwidth.workspace = true
 libc.workspace = true
 once_cell.workspace = true
+widestring.workspace = true
+
+[build-dependencies]
+fish-build-helper.workspace = true
+rsconf.workspace = true
 
 [lints]
 workspace = true
