+++ src/key.rs
@@ -2,13 +2,13 @@ use libc::VERASE;
 
 use crate::{
     common::{EscapeFlags, EscapeStringStyle, escape_string},
-    fallback::fish_wcwidth,
     flog::FloggableDebug,
     future_feature_flags::{FeatureFlag, test as feature_test},
     prelude::*,
     reader::safe_get_terminal_mode_on_startup,
     wutil::fish_wcstoul,
 };
+use fish_fallback::fish_wcwidth;
 use fish_wchar::decode_byte_from_char;
 
 pub(crate) const Backspace: char = '\u{F500}'; // below ENCODE_DIRECT_BASE
